+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Added:
 - App state information
 - Improved Error handling
 - Improved mobile style
+- Auto-discover Servers and Services
 
 Changed:
 - Moved cron execution into the docker image